Conflicts:
	@chebfun2/imag.m
	@chebfun2/length.m
	@chebfun2/mesh.m
	@chebfun2/mldivide.m
	@chebfun2/power.m
	@chebfun2/rdivide.m
	@chebfun2/tanh.m
	@chebfun2/times.m
	@chebfun2/trace.m
	@chebfun2/volt.m
	@separableApprox/feval.m
	@separableApprox/normalizePivots.m
	@separableApprox/normalizeRowsAndCols.m
	@separableApprox/restrict.m
	@separableApprox/simplify.m
	@separableApprox/size.m

+++ @domain/diff.m
@@ -2,9 +2,9 @@ function F = diff(d, varargin)
 %DIFF      Differentiation operator.
 %   This function is deprecated. Use OPERATORBLOCK.DIFF.
 
-% Copyright 2014 by The University of Oxford and The Chebfun Developers.
+% Copyright 2015 by The University of Oxford and The Chebfun Developers.
 % See http://www.chebfun.org/ for Chebfun information.
 
-F = linop( operatorBlock.diff(d, varargin{:}) );
+F = linop( operatorBlock.diff(double(d), varargin{:}) );
 
 end
